dashboard/requirements.txt
requests-oauthlib==1.2.0
flask
Flask-Misaka
pillow
jsonschema
psycopg2-binary